\name{electronics-package}
\alias{electronics-package}
\alias{electronics}
\docType{package}
\title{Electronics}
\description{
This package provides a lot of useful functions for electronics-related applications.
}
\details{
\tabular{ll}{
Package: \tab electronics\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-11\cr
License: \tab MIT License\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Maintainer: Nathan Campos <nathanpc@dreamintech.net>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
